split analyzer2 to multi files
