+++ ci/scripts/intel_copyright_files.txt
@@ -22,7 +22,7 @@ cider/exec/plan/parser/*Visitor*
 cider/exec/plan/parser/Substrait*
 cider/exec/plan/parser/ConverterHelper.*
 cider/exec/nextgen/*
-cider/type/plan/Analyzer2.cpp
+cider/type/plan/*Expr.cpp
 cider/function/FunctionLookupEngine.*
 cider/function/SubstraitFunctionCiderMappings.h
 cider/function/substrait/*
