+++ cider/type/plan/Analyzer.h
@@ -27,14 +27,6 @@
 #ifndef ANALYZER_H
 #define ANALYZER_H
 
-#include "cider/CiderException.h"
-#include "exec/nextgen/jitlib/base/JITFunction.h"
-#include "exec/nextgen/jitlib/base/JITValue.h"
-#include "type/data/sqltypes.h"
-#include "type/schema/ColumnInfo.h"
-#include "util/Logger.h"
-#include "util/sqldefs.h"
-
 #include <cstdint>
 #include <iostream>
 #include <list>
@@ -45,19 +37,14 @@
 #include <utility>
 #include <vector>
 
-namespace Analyzer {
-using namespace cider::jitlib;
-class Expr;
-using ExprPtr = std::shared_ptr<Expr>;
-using ExprPtrRefVector = std::vector<ExprPtr*>;
-}  // namespace Analyzer
-
-template <typename Tp, typename... Args>
-inline typename std::enable_if<std::is_base_of<Analyzer::Expr, Tp>::value,
-                               std::shared_ptr<Tp>>::type
-makeExpr(Args&&... args) {
-  return std::make_shared<Tp>(std::forward<Args>(args)...);
-}
+#include "BinaryExpr.h"
+#include "ColumnExpr.h"
+#include "ConstantExpr.h"
+#include "Expr.h"
+#include "cider/CiderException.h"
+#include "type/data/sqltypes.h"
+#include "util/Logger.h"
+#include "util/sqldefs.h"
 
 enum OpSupportExprType {
   kCOLUMN_VAR,
@@ -96,252 +83,6 @@ enum OpSupportExprType {
 
 namespace Analyzer {
 
-class ColumnVar;
-class TargetEntry;
-class Expr;
-using DomainSet = std::list<const Expr*>;
-
-/*
- * @type Expr
- * @brief super class for all expressions in parse trees and in query plans
- */
-class Expr : public std::enable_shared_from_this<Expr> {
- public:
-  Expr(SQLTypes t, bool notnull) : type_info(t, notnull), contains_agg(false) {}
-  Expr(SQLTypes t, int d, bool notnull)
-      : type_info(t, d, 0, notnull), contains_agg(false) {}
-  Expr(SQLTypes t, int d, int s, bool notnull)
-      : type_info(t, d, s, notnull), contains_agg(false) {}
-  Expr(const SQLTypeInfo& ti, bool has_agg = false)
-      : type_info(ti), contains_agg(has_agg) {}
-  virtual ~Expr() {}
-  std::shared_ptr<Analyzer::Expr> get_shared_ptr() { return shared_from_this(); }
-  const SQLTypeInfo& get_type_info() const { return type_info; }
-  virtual void set_type_info(const SQLTypeInfo& ti) { type_info = ti; }
-  bool get_contains_agg() const { return contains_agg; }
-  void set_contains_agg(bool a) { contains_agg = a; }
-  virtual std::shared_ptr<Analyzer::Expr> add_cast(const SQLTypeInfo& new_type_info);
-  virtual void check_group_by(
-      const std::list<std::shared_ptr<Analyzer::Expr>>& groupby) const {};
-  virtual std::shared_ptr<Analyzer::Expr> deep_copy()
-      const = 0;  // make a deep copy of self
-                  /*
-                   * @brief normalize_simple_predicate only applies to boolean expressions.
-                   * it checks if it is an expression comparing a column
-                   * with a constant.  if so, it returns a normalized copy of the predicate with ColumnVar
-                   * always as the left operand with rte_idx set to the rte_idx of the ColumnVar.
-                   * it returns nullptr with rte_idx set to -1 otherwise.
-                   */
-  virtual std::shared_ptr<Analyzer::Expr> normalize_simple_predicate(int& rte_idx) const {
-    rte_idx = -1;
-    return nullptr;
-  }
-  /*
-   * @brief seperate conjunctive predicates into scan predicates, join predicates and
-   * constant predicates.
-   */
-  virtual void group_predicates(std::list<const Expr*>& scan_predicates,
-                                std::list<const Expr*>& join_predicates,
-                                std::list<const Expr*>& const_predicates) const {}
-  /*
-   * @brief collect_rte_idx collects the indices of all the range table
-   * entries involved in an expression
-   */
-  virtual void collect_rte_idx(std::set<int>& rte_idx_set) const {}
-  /*
-   * @brief collect_column_var collects all unique ColumnVar nodes in an expression
-   * If include_agg = false, it does not include to ColumnVar nodes inside
-   * the argument to AggExpr's.  Otherwise, they are included.
-   * It does not make copies of the ColumnVar
-   */
-  virtual void collect_column_var(
-      std::set<const ColumnVar*, bool (*)(const ColumnVar*, const ColumnVar*)>&
-          colvar_set,
-      bool include_agg) const {}
-  /*
-   * @brief rewrite_with_targetlist rewrite ColumnVar's in expression with entries in a
-   * targetlist. targetlist expressions are expected to be only Var's or AggExpr's.
-   * returns a new expression copy
-   */
-  virtual std::shared_ptr<Analyzer::Expr> rewrite_with_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const {
-    return deep_copy();
-  };
-  /*
-   * @brief rewrite_with_child_targetlist rewrite ColumnVar's in expression with entries
-   * in a child plan's targetlist. targetlist expressions are expected to be only Var's or
-   * ColumnVar's returns a new expression copy
-   */
-  virtual std::shared_ptr<Analyzer::Expr> rewrite_with_child_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const {
-    return deep_copy();
-  };
-  /*
-   * @brief rewrite_agg_to_var rewrite ColumnVar's in expression with entries in an
-   * AggPlan's targetlist. targetlist expressions are expected to be only Var's or
-   * ColumnVar's or AggExpr's All AggExpr's are written into Var's. returns a new
-   * expression copy
-   */
-  virtual std::shared_ptr<Analyzer::Expr> rewrite_agg_to_var(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const {
-    return deep_copy();
-  }
-  virtual bool operator==(const Expr& rhs) const = 0;
-  virtual std::string toString() const = 0;
-  virtual void print() const { std::cout << toString(); }
-
-  virtual void add_unique(std::list<const Expr*>& expr_list) const;
-  /*
-   * @brief find_expr traverse Expr hierarchy and adds the node pointer to
-   * the expr_list if the function f returns true.
-   * Duplicate Expr's are not added the list.
-   * Cannot use std::set because we don't have an ordering function.
-   */
-  virtual void find_expr(bool (*f)(const Expr*),
-                         std::list<const Expr*>& expr_list) const {
-    if (f(this)) {
-      add_unique(expr_list);
-    }
-  }
-  /*
-   * @brief decompress adds cast operator to decompress encoded result
-   */
-  std::shared_ptr<Analyzer::Expr> decompress();
-  /*
-   * @brief perform domain analysis on Expr and fill in domain
-   * information in domain_set.  Empty domain_set means no information.
-   */
-  virtual void get_domain(DomainSet& domain_set) const { domain_set.clear(); }
-
- public:
-  // change this to pure virtual method after all subclasses support codegen.
-  virtual JITExprValue& codegen(JITFunction& func);
-
-  // for vector<JITValuePointer>;
-  template <typename T>
-  JITExprValue& set_expr_value(T&& ptrs, bool is_variadic = false) {
-    expr_var_ = std::make_unique<JITExprValue>(std::forward<T>(ptrs), is_variadic);
-    return *expr_var_;
-  }
-
-  // for {JITValuePointer, ...}
-  template <typename... T>
-  JITExprValue& set_expr_value(T&&... ptrs, bool is_variadic = false) {
-    expr_var_ = std::make_unique<JITExprValue>(std::forward<T>(ptrs)...);
-    return *expr_var_;
-  }
-
-  // for JITValuePointer
-  // used with only value (no len and null, so is_variadic is false)
-  JITExprValue& set_expr_value(JITValuePointer&& val) {
-    expr_var_ = std::make_unique<JITExprValue>(std::move(val));
-    return *expr_var_;
-  }
-
-  JITExprValue* get_expr_value() const { return expr_var_.get(); }
-
-  // TODO (bigPYJ1151): to pure virtual.
-  virtual ExprPtrRefVector get_children_reference() {
-    UNREACHABLE();
-    return {};
-  }
-
- protected:
-  JITTypeTag getJITTag(const SQLTypes& st);
-  JITTypeTag getJITTag() { return getJITTag(get_type_info().get_type()); }
-
- protected:
-  SQLTypeInfo type_info;  // SQLTypeInfo of the return result of this expression
-  bool contains_agg;
-
-  std::unique_ptr<JITExprValue> expr_var_;
-  // just for unreachable branch return;
-  JITExprValue fake_val_;
-};
-
-using ExpressionPtr = std::shared_ptr<Analyzer::Expr>;
-using ExpressionPtrList = std::list<ExpressionPtr>;
-using ExpressionPtrVector = std::vector<ExpressionPtr>;
-
-/*
- * @type ColumnVar
- * @brief expression that evaluates to the value of a column in a given row from a base
- * table. It is used in parse trees and is only used in Scan nodes in a query plan for
- * scanning a table while Var nodes are used for all other plans.
- */
-class ColumnVar : public Expr {
- public:
-  ColumnVar(ColumnInfoPtr col_info, int nest_level)
-      : Expr(col_info->type), rte_idx(nest_level), col_info_(std::move(col_info)) {}
-  ColumnVar(const SQLTypeInfo& ti)
-      : Expr(ti)
-      , rte_idx(-1)
-      , col_info_(std::make_shared<ColumnInfo>(-1, 0, 0, "", ti, false)) {}
-  ColumnVar(const SQLTypeInfo& ti,
-            int table_id,
-            int col_id,
-            int nest_level,
-            bool is_virtual = false)
-      : Expr(ti)
-      , rte_idx(nest_level)
-      , col_info_(
-            std::make_shared<ColumnInfo>(-1, table_id, col_id, "", ti, is_virtual)) {}
-  int get_db_id() const { return col_info_->db_id; }
-  int get_table_id() const { return col_info_->table_id; }
-  int get_column_id() const { return col_info_->column_id; }
-  int get_rte_idx() const { return rte_idx; }
-  ColumnInfoPtr get_column_info() const { return col_info_; }
-  bool is_virtual() const { return col_info_->is_rowid; }
-  EncodingType get_compression() const { return type_info.get_compression(); }
-  int get_comp_param() const { return type_info.get_comp_param(); }
-  void set_type_info(const SQLTypeInfo& ti) override {
-    if (type_info != ti) {
-      col_info_ = std::make_shared<ColumnInfo>(col_info_->db_id,
-                                               col_info_->table_id,
-                                               col_info_->column_id,
-                                               col_info_->name,
-                                               ti,
-                                               col_info_->is_rowid);
-      type_info = ti;
-    }
-  }
-  void check_group_by(
-      const std::list<std::shared_ptr<Analyzer::Expr>>& groupby) const override;
-  std::shared_ptr<Analyzer::Expr> deep_copy() const override;
-  void group_predicates(std::list<const Expr*>& scan_predicates,
-                        std::list<const Expr*>& join_predicates,
-                        std::list<const Expr*>& const_predicates) const override;
-  void collect_rte_idx(std::set<int>& rte_idx_set) const override {
-    rte_idx_set.insert(rte_idx);
-  }
-  static bool colvar_comp(const ColumnVar* l, const ColumnVar* r) {
-    return l->get_table_id() < r->get_table_id() ||
-           (l->get_table_id() == r->get_table_id() &&
-            l->get_column_id() < r->get_column_id());
-  }
-  void collect_column_var(
-      std::set<const ColumnVar*, bool (*)(const ColumnVar*, const ColumnVar*)>&
-          colvar_set,
-      bool include_agg) const override {
-    colvar_set.insert(this);
-  }
-  std::shared_ptr<Analyzer::Expr> rewrite_with_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override;
-  std::shared_ptr<Analyzer::Expr> rewrite_with_child_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override;
-  std::shared_ptr<Analyzer::Expr> rewrite_agg_to_var(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override;
-  bool operator==(const Expr& rhs) const override;
-  std::string toString() const override;
-
- public:
-  JITExprValue& codegen(JITFunction& func) override;
-
- protected:
-  int rte_idx;  // 0-based range table index, used for table ordering in multi-joins
-  ColumnInfoPtr col_info_;
-};
-
 /*
  * @type ExpressionTuple
  * @brief A tuple of expressions on the side of an equi-join on multiple columns.
@@ -363,110 +104,6 @@ class ExpressionTuple : public Expr {
 
  private:
   const std::vector<std::shared_ptr<Analyzer::Expr>> tuple_;
-};
-
-/*
- * @type Var
- * @brief expression that evaluates to the value of a column in a given row generated
- * from a query plan node.  It is only used in plan nodes above Scan nodes.
- * The row can be produced by either the inner or the outer plan in case of a join.
- * It inherits from ColumnVar to keep track of the lineage through the plan nodes.
- * The table_id will be set to 0 if the Var does not correspond to an original column
- * value.
- */
-class Var : public ColumnVar {
- public:
-  enum WhichRow { kINPUT_OUTER, kINPUT_INNER, kOUTPUT, kGROUPBY };
-  Var(const SQLTypeInfo& ti, int r, int c, int i, bool is_virtual, WhichRow o, int v)
-      : ColumnVar(ti, r, c, i, is_virtual), which_row(o), varno(v) {}
-  Var(ColumnInfoPtr col_info, int i, WhichRow o, int v)
-      : ColumnVar(col_info, i), which_row(o), varno(v) {}
-  Var(const SQLTypeInfo& ti, WhichRow o, int v) : ColumnVar(ti), which_row(o), varno(v) {}
-  WhichRow get_which_row() const { return which_row; }
-  void set_which_row(WhichRow r) { which_row = r; }
-  int get_varno() const { return varno; }
-  void set_varno(int n) { varno = n; }
-  std::shared_ptr<Analyzer::Expr> deep_copy() const override;
-  std::string toString() const override;
-  void check_group_by(
-      const std::list<std::shared_ptr<Analyzer::Expr>>& groupby) const override;
-  void collect_rte_idx(std::set<int>& rte_idx_set) const override {
-    rte_idx_set.insert(-1);
-  }
-  std::shared_ptr<Analyzer::Expr> rewrite_with_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
-    return deep_copy();
-  }
-  std::shared_ptr<Analyzer::Expr> rewrite_with_child_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
-    return deep_copy();
-  }
-  std::shared_ptr<Analyzer::Expr> rewrite_agg_to_var(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override;
-
- private:
-  WhichRow which_row;  // indicate which row this Var should project from.  It can be from
-                       // the outer input plan or the inner input plan (for joins) or the
-                       // output row in the current plan.
-  int varno;           // the column number in the row.  1-based
-};
-
-/*
- * @type Constant
- * @brief expression for a constant value
- */
-class Constant : public Expr {
- public:
-  Constant(SQLTypes t, bool n) : Expr(t, !n), is_null(n) {
-    if (n) {
-      set_null_value();
-    } else {
-      type_info.set_notnull(true);
-    }
-  }
-  Constant(SQLTypes t, bool n, Datum v) : Expr(t, !n), is_null(n), constval(v) {
-    if (n) {
-      set_null_value();
-    } else {
-      type_info.set_notnull(true);
-    }
-  }
-  Constant(const SQLTypeInfo& ti, bool n, Datum v) : Expr(ti), is_null(n), constval(v) {
-    if (n) {
-      set_null_value();
-    } else {
-      type_info.set_notnull(true);
-    }
-  }
-  Constant(const SQLTypeInfo& ti,
-           bool n,
-           const std::list<std::shared_ptr<Analyzer::Expr>>& l)
-      : Expr(ti), is_null(n), constval(Datum{0}), value_list(l) {}
-  ~Constant() override;
-  bool get_is_null() const { return is_null; }
-  Datum get_constval() const { return constval; }
-  void set_constval(Datum d) { constval = d; }
-  const std::list<std::shared_ptr<Analyzer::Expr>>& get_value_list() const {
-    return value_list;
-  }
-  std::shared_ptr<Analyzer::Expr> deep_copy() const override;
-  std::shared_ptr<Analyzer::Expr> add_cast(const SQLTypeInfo& new_type_info) override;
-  bool operator==(const Expr& rhs) const override;
-  std::string toString() const override;
-
- public:
-  JITExprValue& codegen(JITFunction& func) override;
-
- protected:
-  bool is_null;    // constant is NULL
-  Datum constval;  // the constant value
-  const std::list<std::shared_ptr<Analyzer::Expr>> value_list;
-  void cast_number(const SQLTypeInfo& new_type_info);
-  void cast_string(const SQLTypeInfo& new_type_info);
-  void cast_from_string(const SQLTypeInfo& new_type_info);
-  void cast_to_string(const SQLTypeInfo& new_type_info);
-  void do_cast(const SQLTypeInfo& new_type_info);
-  void set_null_value();
 };
 
 /*
@@ -525,125 +162,6 @@ class UOper : public Expr {
  protected:
   SQLOps optype;  // operator type, e.g., kUMINUS, kISNULL, kEXISTS
   std::shared_ptr<Analyzer::Expr> operand;  // operand expression
-};
-
-/*
- * @type BinOper
- * @brief represents binary operator expressions.  it includes all
- * comparison, arithmetic and boolean binary operators.  it handles ANY/ALL qualifiers
- * in case the right_operand is a subquery.
- */
-class BinOper : public Expr {
- public:
-  BinOper(const SQLTypeInfo& ti,
-          bool has_agg,
-          SQLOps o,
-          SQLQualifier q,
-          std::shared_ptr<Analyzer::Expr> l,
-          std::shared_ptr<Analyzer::Expr> r)
-      : Expr(ti, has_agg), optype(o), qualifier(q), left_operand(l), right_operand(r) {}
-  BinOper(SQLTypes t,
-          SQLOps o,
-          SQLQualifier q,
-          std::shared_ptr<Analyzer::Expr> l,
-          std::shared_ptr<Analyzer::Expr> r)
-      : Expr(t, l->get_type_info().get_notnull() && r->get_type_info().get_notnull())
-      , optype(o)
-      , qualifier(q)
-      , left_operand(l)
-      , right_operand(r) {}
-  SQLOps get_optype() const { return optype; }
-  SQLQualifier get_qualifier() const { return qualifier; }
-  const Expr* get_left_operand() const { return left_operand.get(); }
-  const Expr* get_right_operand() const { return right_operand.get(); }
-  std::shared_ptr<Analyzer::Expr> get_non_const_left_operand() const {
-    return left_operand;
-  }
-  const std::shared_ptr<Analyzer::Expr> get_own_left_operand() const {
-    return left_operand;
-  }
-  const std::shared_ptr<Analyzer::Expr> get_own_right_operand() const {
-    return right_operand;
-  }
-  std::shared_ptr<Analyzer::Expr> get_non_const_right_operand() const {
-    return right_operand;
-  }
-  static SQLTypeInfo analyze_type_info(SQLOps op,
-                                       const SQLTypeInfo& left_type,
-                                       const SQLTypeInfo& right_type,
-                                       SQLTypeInfo* new_left_type,
-                                       SQLTypeInfo* new_right_type);
-  static SQLTypeInfo common_numeric_type(const SQLTypeInfo& type1,
-                                         const SQLTypeInfo& type2);
-  static SQLTypeInfo common_string_type(const SQLTypeInfo& type1,
-                                        const SQLTypeInfo& type2);
-  void check_group_by(
-      const std::list<std::shared_ptr<Analyzer::Expr>>& groupby) const override;
-  std::shared_ptr<Analyzer::Expr> deep_copy() const override;
-  std::shared_ptr<Analyzer::Expr> normalize_simple_predicate(int& rte_idx) const override;
-  void group_predicates(std::list<const Expr*>& scan_predicates,
-                        std::list<const Expr*>& join_predicates,
-                        std::list<const Expr*>& const_predicates) const override;
-  void collect_rte_idx(std::set<int>& rte_idx_set) const override {
-    left_operand->collect_rte_idx(rte_idx_set);
-    right_operand->collect_rte_idx(rte_idx_set);
-  }
-  void collect_column_var(
-      std::set<const ColumnVar*, bool (*)(const ColumnVar*, const ColumnVar*)>&
-          colvar_set,
-      bool include_agg) const override {
-    left_operand->collect_column_var(colvar_set, include_agg);
-    right_operand->collect_column_var(colvar_set, include_agg);
-  }
-  std::shared_ptr<Analyzer::Expr> rewrite_with_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
-    return makeExpr<BinOper>(type_info,
-                             contains_agg,
-                             optype,
-                             qualifier,
-                             left_operand->rewrite_with_targetlist(tlist),
-                             right_operand->rewrite_with_targetlist(tlist));
-  }
-  std::shared_ptr<Analyzer::Expr> rewrite_with_child_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
-    return makeExpr<BinOper>(type_info,
-                             contains_agg,
-                             optype,
-                             qualifier,
-                             left_operand->rewrite_with_child_targetlist(tlist),
-                             right_operand->rewrite_with_child_targetlist(tlist));
-  }
-  std::shared_ptr<Analyzer::Expr> rewrite_agg_to_var(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
-    return makeExpr<BinOper>(type_info,
-                             contains_agg,
-                             optype,
-                             qualifier,
-                             left_operand->rewrite_agg_to_var(tlist),
-                             right_operand->rewrite_agg_to_var(tlist));
-  }
-  bool operator==(const Expr& rhs) const override;
-  std::string toString() const override;
-  void find_expr(bool (*f)(const Expr*),
-                 std::list<const Expr*>& expr_list) const override;
-  static bool simple_predicate_has_simple_cast(
-      const std::shared_ptr<Analyzer::Expr> cast_operand,
-      const std::shared_ptr<Analyzer::Expr> const_operand);
-
- public:
-  ExprPtrRefVector get_children_reference() override {
-    return {&left_operand, &right_operand};
-  }
-  JITExprValue& codegen(JITFunction& func) override;
-  JITExprValue& codegenFixedSizeColArithFun(JITValue& lhs, JITValue& rhs);
-  JITExprValue& codegenFixedSizeColCmpFun(JITValue& lhs, JITValue& rhs);
-
- private:
-  SQLOps optype;           // operator type, e.g., kLT, kAND, kPLUS, etc.
-  SQLQualifier qualifier;  // qualifier kANY, kALL or kONE.  Only relevant with
-                           // right_operand is Subquery
-  std::shared_ptr<Analyzer::Expr> left_operand;   // the left operand expression
-  std::shared_ptr<Analyzer::Expr> right_operand;  // the right operand expression
 };
 
 /**
