+++ cider/type/plan/BinaryExpr.h
@@ -1,0 +1,158 @@
+/*
+ * Copyright (c) 2022 Intel Corporation.
+ * Copyright (c) OmniSci, Inc. and its affiliates.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+#ifndef TYPE_PLAN_BINARY_EXPR_H
+#define TYPE_PLAN_BINARY_EXPR_H
+
+#include <list>
+#include <memory>
+#include <set>
+#include <string>
+#include <vector>
+
+#include "Expr.h"
+#include "exec/nextgen/jitlib/JITLib.h"
+#include "type/data/sqltypes.h"
+#include "util/sqldefs.h"
+
+namespace Analyzer {
+
+/*
+ * @type BinOper
+ * @brief represents binary operator expressions.  it includes all
+ * comparison, arithmetic and boolean binary operators.  it handles ANY/ALL qualifiers
+ * in case the right_operand is a subquery.
+ */
+class BinOper : public Expr {
+ public:
+  BinOper(const SQLTypeInfo& ti,
+          bool has_agg,
+          SQLOps o,
+          SQLQualifier q,
+          std::shared_ptr<Analyzer::Expr> l,
+          std::shared_ptr<Analyzer::Expr> r)
+      : Expr(ti, has_agg), optype(o), qualifier(q), left_operand(l), right_operand(r) {}
+  BinOper(SQLTypes t,
+          SQLOps o,
+          SQLQualifier q,
+          std::shared_ptr<Analyzer::Expr> l,
+          std::shared_ptr<Analyzer::Expr> r)
+      : Expr(t, l->get_type_info().get_notnull() && r->get_type_info().get_notnull())
+      , optype(o)
+      , qualifier(q)
+      , left_operand(l)
+      , right_operand(r) {}
+  SQLOps get_optype() const { return optype; }
+  SQLQualifier get_qualifier() const { return qualifier; }
+  const Expr* get_left_operand() const { return left_operand.get(); }
+  const Expr* get_right_operand() const { return right_operand.get(); }
+  std::shared_ptr<Analyzer::Expr> get_non_const_left_operand() const {
+    return left_operand;
+  }
+  const std::shared_ptr<Analyzer::Expr> get_own_left_operand() const {
+    return left_operand;
+  }
+  const std::shared_ptr<Analyzer::Expr> get_own_right_operand() const {
+    return right_operand;
+  }
+  std::shared_ptr<Analyzer::Expr> get_non_const_right_operand() const {
+    return right_operand;
+  }
+  static SQLTypeInfo analyze_type_info(SQLOps op,
+                                       const SQLTypeInfo& left_type,
+                                       const SQLTypeInfo& right_type,
+                                       SQLTypeInfo* new_left_type,
+                                       SQLTypeInfo* new_right_type);
+  static SQLTypeInfo common_numeric_type(const SQLTypeInfo& type1,
+                                         const SQLTypeInfo& type2);
+  static SQLTypeInfo common_string_type(const SQLTypeInfo& type1,
+                                        const SQLTypeInfo& type2);
+  void check_group_by(
+      const std::list<std::shared_ptr<Analyzer::Expr>>& groupby) const override;
+  std::shared_ptr<Analyzer::Expr> deep_copy() const override;
+  std::shared_ptr<Analyzer::Expr> normalize_simple_predicate(int& rte_idx) const override;
+  void group_predicates(std::list<const Expr*>& scan_predicates,
+                        std::list<const Expr*>& join_predicates,
+                        std::list<const Expr*>& const_predicates) const override;
+  void collect_rte_idx(std::set<int>& rte_idx_set) const override {
+    left_operand->collect_rte_idx(rte_idx_set);
+    right_operand->collect_rte_idx(rte_idx_set);
+  }
+  void collect_column_var(
+      std::set<const ColumnVar*, bool (*)(const ColumnVar*, const ColumnVar*)>&
+          colvar_set,
+      bool include_agg) const override {
+    left_operand->collect_column_var(colvar_set, include_agg);
+    right_operand->collect_column_var(colvar_set, include_agg);
+  }
+  std::shared_ptr<Analyzer::Expr> rewrite_with_targetlist(
+      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
+    return makeExpr<BinOper>(type_info,
+                             contains_agg,
+                             optype,
+                             qualifier,
+                             left_operand->rewrite_with_targetlist(tlist),
+                             right_operand->rewrite_with_targetlist(tlist));
+  }
+  std::shared_ptr<Analyzer::Expr> rewrite_with_child_targetlist(
+      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
+    return makeExpr<BinOper>(type_info,
+                             contains_agg,
+                             optype,
+                             qualifier,
+                             left_operand->rewrite_with_child_targetlist(tlist),
+                             right_operand->rewrite_with_child_targetlist(tlist));
+  }
+  std::shared_ptr<Analyzer::Expr> rewrite_agg_to_var(
+      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override {
+    return makeExpr<BinOper>(type_info,
+                             contains_agg,
+                             optype,
+                             qualifier,
+                             left_operand->rewrite_agg_to_var(tlist),
+                             right_operand->rewrite_agg_to_var(tlist));
+  }
+  bool operator==(const Expr& rhs) const override;
+  std::string toString() const override;
+  void find_expr(bool (*f)(const Expr*),
+                 std::list<const Expr*>& expr_list) const override;
+  static bool simple_predicate_has_simple_cast(
+      const std::shared_ptr<Analyzer::Expr> cast_operand,
+      const std::shared_ptr<Analyzer::Expr> const_operand);
+
+ public:
+  ExprPtrRefVector get_children_reference() override {
+    return {&left_operand, &right_operand};
+  }
+  JITExprValue& codegen(JITFunction& func) override;
+  JITExprValue& codegenFixedSizeColArithFun(JITValue& lhs, JITValue& rhs);
+  JITExprValue& codegenFixedSizeColCmpFun(JITValue& lhs, JITValue& rhs);
+
+ private:
+  SQLOps optype;           // operator type, e.g., kLT, kAND, kPLUS, etc.
+  SQLQualifier qualifier;  // qualifier kANY, kALL or kONE.  Only relevant with
+                           // right_operand is Subquery
+  std::shared_ptr<Analyzer::Expr> left_operand;   // the left operand expression
+  std::shared_ptr<Analyzer::Expr> right_operand;  // the right operand expression
+};
+}  // namespace Analyzer
+
+#endif
