+++ cider/type/plan/CMakeLists.txt
@@ -17,5 +17,6 @@
 # specific language governing permissions and limitations
 # under the License.
 
-set(cider_type_plan_source_files Analyzer.cpp Analyzer2.cpp)
+set(cider_type_plan_source_files Analyzer.cpp Expr.cpp BinaryExpr.cpp
+                                 ColumnExpr.cpp ConstantExpr.cpp)
 add_library(cider_type_plan ${cider_type_plan_source_files})
