+++ cider/type/plan/Expr.cpp
@@ -1,0 +1,62 @@
+
+/*
+ * Copyright (c) 2022 Intel Corporation.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+#include "Expr.h"
+
+namespace Analyzer {
+using namespace cider::jitlib;
+
+// change this to pure virtual method after all subclasses support codegen.
+JITExprValue& Expr::codegen(JITFunction& func) {
+  UNREACHABLE();
+  return fake_val_;
+}
+
+JITTypeTag Expr::getJITTag(const SQLTypes& st) {
+  switch (st) {
+    case kBOOLEAN:
+      return JITTypeTag::BOOL;
+    case kTINYINT:
+    case kSMALLINT:
+    case kINT:
+    case kBIGINT:
+    case kTIME:
+    case kTIMESTAMP:
+    case kDATE:
+    case kINTERVAL_DAY_TIME:
+    case kINTERVAL_YEAR_MONTH:
+      return JITTypeTag::INT32;
+    case kFLOAT:
+      return JITTypeTag::FLOAT;
+    case kDOUBLE:
+      return JITTypeTag::DOUBLE;
+    case kVARCHAR:
+    case kCHAR:
+    case kTEXT:
+      UNIMPLEMENTED();
+    case kNULLT:
+    default:
+      return JITTypeTag::INVALID;
+  }
+  UNREACHABLE();
+}
+
+}  // namespace Analyzer
